requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
astor==0.8.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.2.0
Babel==2.11.0
bce-python-sdk==0.8.79
beautifulsoup4==4.11.2
Brotli==1.0.9
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
dill==0.3.6
einops==0.6.0
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.92.0
ffmpy==0.3.0
filelock==3.9.0
fire==0.5.0
Flask==2.2.2
flask-babel==3.0.1
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gdown==4.6.3
gevent==22.10.2
geventhttpclient==2.0.2
gradio==3.18.0
greenlet==2.0.2
grpcio==1.51.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
imageio==2.25.1
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==1.0.3
lmdb==1.4.0
lxml==4.9.2
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
onnx==1.13.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.1
opt-einsum==3.3.0
orjson==3.8.6
packaging==23.0
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.4.1
pandas==1.5.3
pdf2docx==0.5.6
Pillow==9.4.0
premailer==3.10.0
protobuf==3.20.0
psutil==5.9.4
pyclipper==1.3.0.post4
pycryptodome==3.17
pydantic==1.10.4
pydub==0.25.1
PyMuPDF==1.20.2
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-multipart==0.0.5
python-rapidjson==1.9
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
rank-bm25==0.2.2
rapidfuzz==2.13.7
rarfile==4.0
requests==2.28.2
rfc3986==1.5.0
scikit-image==0.19.3
scipy==1.10.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.25.0
sympy==1.11.1
termcolor==2.2.0
tifffile==2023.2.3
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
tritonclient==2.30.0
typing-extensions==4.5.0
uc-micro-py==1.0.1
urllib3==1.26.14
uvicorn==0.20.0
visualdl==2.5.0
websockets==10.4
Werkzeug==2.2.3
x2paddle==1.4.0
yarl==1.8.2
zipp==3.13.0
zope.event==4.6
zope.interface==5.5.2
